+++ src/main/java/com/pewpew/pewpew/common/Validate.java
@@ -1,4 +1,4 @@
-package com.pewpew.pewpew.additional;
+package com.pewpew.pewpew.common;
 
 import com.pewpew.pewpew.model.User;
 import org.jetbrains.annotations.NotNull;
@@ -10,18 +10,18 @@ public class Validate {
     public static boolean userRegister(User user) {
         if (user.getEmail() == null) return false;
         if (user.getLogin() == null) return false;
+        if (user.getPassword() == null) return false;
         if (user.getEmail().isEmpty()) return false;
         if (user.getLogin().isEmpty()) return false;
-        if (user.getPassword().isEmpty()) return false;
-        return user.getPassword() != null;
+        return !(user.getPassword().isEmpty());
     }
 
     @NotNull
     public static boolean userAuth(User user) {
         if (user.getLogin() == null) return false;
+        if (user.getPassword() == null) return false;
         if (user.getLogin().isEmpty()) return false;
-        if (user.getPassword().isEmpty()) return false;
-        return user.getPassword() != null;
+        return !(user.getPassword().isEmpty());
     }
 
     @NotNull
